+++ packages/vsce/__tests__/__e2e__/specs/tree.spec.ts
@@ -1,0 +1,59 @@
+import { expect, test } from "@playwright/test";
+import { constants, findAndClickTreeItem, getTreeItem, prepareZoweExplorerView, resetWiremock, resetZoweExplorerView } from "../utils/helpers";
+
+test.beforeEach(async ({ page, request }) => {
+  await resetWiremock(request);
+  await prepareZoweExplorerView(page);
+});
+
+test.afterEach(async ({ page }) => {
+  await resetZoweExplorerView(page);
+});
+
+test("CICSPlexTree is organised in correct order", async ({ page }) => {
+  const expectedOrder = [
+    "Bundles",
+    "Files",
+    "JVM Servers",
+    "Libraries",
+    "Pipelines",
+    "Programs",
+    "Tasks",
+    "TCP/IP Services",
+    "Transactions",
+    "TS Queues",
+    "URI Maps",
+    "Web Services",
+  ];
+
+  await findAndClickTreeItem(page, constants.PROFILE_NAME);
+  await findAndClickTreeItem(page, constants.CICSPLEX_NAME);
+  await findAndClickTreeItem(page, constants.REGION_NAME);
+  for (const order of expectedOrder) {
+    await expect(getTreeItem(page, order)).toBeVisible();
+  }
+});
+
+test("CICSRegionTree is organised in correct order", async ({ page }) => {
+  const expectedOrder = [
+    "Regions",
+    "All Bundles",
+    "All Files",
+    "All JVM Servers",
+    "All Libraries",
+    "All Pipelines",
+    "All Programs",
+    "All Tasks",
+    "All TCP/IP Services",
+    "All Transactions",
+    "All TS Queues",
+    "All URI Maps",
+    "All Web Services",
+  ];
+
+  await findAndClickTreeItem(page, constants.PROFILE_NAME);
+  await findAndClickTreeItem(page, constants.CICSPLEX_NAME);
+  for (const order of expectedOrder) {
+    await expect(getTreeItem(page, order)).toBeVisible();
+  }
+});